+++ examples/async_http_array_read.rs
@@ -1,4 +1,3 @@
-#[cfg(all(feature = "ndarray", feature = "async"))]
 async fn http_array_read() -> Result<(), Box<dyn std::error::Error>> {
     use std::sync::Arc;
     use zarrs::{
@@ -56,11 +55,6 @@ async fn http_array_read() -> Result<(), Box<dyn std::error::Error>> {
     println!("The middle 4x2 subset is:\n{:?}\n", data_4x2);
 
     Ok(())
-}
-
-#[cfg(any(not(feature = "ndarray"), not(feature = "async")))]
-async fn http_array_read() -> Result<(), Box<dyn std::error::Error>> {
-    panic!("the async_http_array_read example requires the ndarray and async features")
 }
 
 #[tokio::main]